refactor(datapack): 매니페스트 검증을 lib/manifest-validation로 추출 (#1691)
